Fixing delete bug

+++ src/org/pahappa/systems/ticketing/views/TicketView.java
@@ -232,7 +232,8 @@ public class TicketView implements BaseTicketView {
 
     @Override
     public void deleteTicket() {
-        System.out.println("Enter ticket Number to delete: ");
+        try{
+            System.out.println("Enter ticket Number to delete: ");
         Scanner sc3 = new Scanner(System.in);
         String input3 = sc3.nextLine();
 
@@ -249,19 +250,21 @@ public class TicketView implements BaseTicketView {
                 System.out.println("STATUS: "+uticket.getStatus());
                 System.out.println("\n");  
 
-                int index= ticketDelete.indexOf(uticket);
-                ticketService.deleteTicket(index);
+                System.out.print("Confirm delete(Enter 2): ");
+                if(sc3.nextInt()==2){
+                   int index = ticketDelete.indexOf(uticket);
+                     ticketService.deleteTicket(index);
+                 }else {
+                    continue;
+                }
+                // int index= ticketDelete.indexOf(uticket);
+                // ticketService.deleteTicket(index);
 
             }
         }
-        // for(Ticket ticket: ticketDelete){
-        //     System.out.print("Confirm delete(Enter 2): ");
-        // if(sc3.nextInt()==2){
-        //     int index = ticketDelete.indexOf(ticket);
-        //     ticketService.deleteTicket(index);
-        // }else {
-        //     continue;
-        //}
+        //  
+        }catch(Exception e){}
+        
         
            
     }
